seems like solved profile URL issue
