+++ src/components/ui/LeftMenu/LeftMenu.tsx
@@ -7,10 +7,9 @@ const LeftMenu: FC<{
   links: IMenuLink[];
   initialStep?: number;
   topContent?: ReactNode;
-  changeParams?: (section: string) => void;
   router: any;
-}> = ({ links, initialStep, topContent, changeParams, router }) => {
-  const [current, setCurrent] = useState(initialStep || 0);
+}> = ({ links, initialStep, topContent, router }) => {
+  // const [current, setCurrent] = useState(initialStep || 0);
   const [initialLoad, setInitialLoad] = useState(true);
   let section = router.query.section as string;
   useEffect(() => {
@@ -19,7 +18,34 @@ const LeftMenu: FC<{
       changeParams!(section!);
       setInitialLoad(false);
     }
+    if (!router.query.section) {
+      changeParams!(links[0].section!);
+    }
   });
+
+  const displayPage = () => {
+    if (router.query.section) {
+      if (
+        links.filter((element) => element.section == router.query.section)[0]
+      ) {
+        return links.filter(
+          (element) => element.section == router.query.section
+        )[0].page;
+      } else {
+        return links[0].page;
+      }
+    } else {
+      return links[0].page;
+    }
+  };
+
+  const changeParams = (section: string) => {
+    const newPathObject = {
+      pathname: router.pathname,
+      query: { section: section },
+    };
+    router.push(newPathObject, undefined, { shallow: true });
+  };
 
   return (
     <div className={styles.wrapper}>
@@ -29,9 +55,9 @@ const LeftMenu: FC<{
           {links.map((element, idx) => (
             <NavLink
               key={idx}
-              active={idx == current}
+              active={element.section == router.query.section}
               onClick={() => {
-                setCurrent(idx);
+                // setCurrent(idx);
                 // changeParams!(idx);
                 changeParams!(links[idx].section!);
               }}
@@ -42,7 +68,8 @@ const LeftMenu: FC<{
         </>
       </Box>
       <div className={styles.pageWrapper}>
-        {links[current]?.page || links[0]?.page || ''}
+        {/* {links[current]?.page || links[0]?.page || ''} */}
+        {displayPage()}
       </div>
     </div>
   );
